+++ example/a.ts
@@ -1,24 +1,18 @@
 import { Elysia } from '../src'
 
-const a = new Elysia()
-	.trace(async ({ set, request }) => {
-		console.log("A")
-		// await request
-		console.log("B")
-	})
+const a = new Elysia({ name: 'p' }).trace(async ({ set, request }) => {
+	console.log('A')
+	// await request
+	console.log('B')
+})
+
+const b = new Elysia({ scoped: true }).get('/scoped', () => 'hi')
 
 const app = new Elysia()
 	.use(a)
-	// .trace(async ({ set, request }) => {
-	// 	console.log("A")
-	// 	// await request
-	// 	console.log("B")
-	// })
-	.get('/', () => {
-		console.log("A")
-
-		return 'hi'
-	})
+	.use(b)
+	.get('/', () => 'hi')
 	.listen(3000)
 
-console.log(app.routes[0].composed?.toString())
+// console.log(app.event)
+// console.log(app.routes[0].composed?.toString())