+++ test/units/utils.ts
@@ -1,0 +1,7 @@
+export class Passthrough {
+    $passthrough = 'custom'
+
+    get custom() {
+        return 'hi'
+    }
+}